Rename shared => types
