+++ addons/docs/src/blocks/Stories.tsx
@@ -3,7 +3,7 @@ import { DocsContext } from './DocsContext';
 import { DocsStory } from './DocsStory';
 import { Heading } from './Heading';
 import { getDocsStories } from './utils';
-import { DocsStoryProps } from './shared';
+import { DocsStoryProps } from './types';
 
 interface StoriesProps {
   title?: JSX.Element | string;
